third_party/blink/web_tests/platform/mac-mac10.15/external/wpt/html/browsers/windows/clear-window-name.https-expected.txt

This is a testharness.js-based test.
PASS Window.name is not reset when there is an opener around
PASS Window.name is not reset when there is an opener around (cross-origin)
PASS Window.name is not reset at the first navigation away from initial about:blank with noopener
FAIL Window.name is reset at the first cross-origin navigation with noopener assert_equals: expected "" but got "af835388-79e2-4051-920c-00689786c08b"
PASS Window.name is not reset at the first navigation away from initial about:blank with window.opener set to null
PASS Window.name is not reset at the same-origin navigation with window.opener set to null
FAIL Window.name is reset at the first first cross-origin navigation with window.opener set to null assert_equals: expected "" but got "48d9612c-e122-4cae-9ef2-03ae3386015b"
PASS Window.name is set by the window
FAIL Window.name is reset at the first cross-origin navigation assert_equals: expected "" but got "008dc857-24e2-4fdf-ac54-083d8847c04c"
Harness: the test ran to completion.
